requirements.txt
flask
pandas
fastapi
uvicorn
Jinja2==3.0.3
fastapi-sqlalchemy
python-dotenv
python-multipart
psycopg2
